wip
alternative implementations

Partitioned Casper integration

wip

+++ block-storage/src/main/scala/coop/rchain/blockstorage/dag/codecs.scala
@@ -35,7 +35,7 @@ object codecs {
 
   val codecValidator = xmapToByteString(bytes(Validator.Length))
 
-  val codecSeqNum = int32
+  val codecSeqNum = int64
 
   val codecBlockHashSet = listOfN(int32, codecBlockHash).xmap[Set[BlockHash]](_.toSet, _.toList)
