+++ node/src/main/scala/coop/rchain/node/api/WebApi.scala
@@ -234,7 +234,7 @@ object WebApi {
 
   final case class PrepareResponse(
       names: List[String],
-      seqNumber: Int
+      seqNumber: Long
   )
 
   final case class ApiStatus(
